This is the beta version of the gear shift program.
This RobotC file is an earlier version of the gear shift program. This beta version does not have any automatic mechanical correction. Thus, the gear shift’s microcontroller “assumes” that the gear’s position is in the correct state upon initialization. If the gear shift is in an intermittent/mechanical error state, the mechanical error will either stay the same or worsen. 
The final version of this program has this issue resolved with automatic start-up detection and correction and automatic general correction.